FAC-406: UseCase moved from space to spaceuseraccess
