+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/adapter/in/rest/spaceuseraccess/UpdateSpaceLastSeenController.java
@@ -1,9 +1,10 @@
-package org.flickit.assessment.users.adapter.in.rest.space;
+package org.flickit.assessment.users.adapter.in.rest.spaceuseraccess;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.users.application.port.in.space.UpdateSpaceLastSeenUseCase;
-import org.flickit.assessment.users.application.port.in.space.UpdateSpaceLastSeenUseCase.Param;
+import org.flickit.assessment.users.adapter.in.rest.space.CreateSpaceResponseDto;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.UpdateSpaceLastSeenUseCase;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.UpdateSpaceLastSeenUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,13 +21,13 @@ public class UpdateSpaceLastSeenController {
     private final UserContext userContext;
 
     @PutMapping("/spaces/{id}/update-last-seen")
-    public ResponseEntity<CreateSpaceResponseDto> updateSpaceLastSeen(@PathVariable("id") Long id) {
+    public ResponseEntity<CreateSpaceResponseDto> updateSpaceLastSeen(@PathVariable("id") Long spaceId) {
         UUID currentUserId = userContext.getUser().id();
-        useCase.updateLastSeen(toParam(id, currentUserId));
+        useCase.updateLastSeen(toParam(spaceId, currentUserId));
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    private Param toParam(Long id, UUID currentUserId) {
-        return new Param(id, currentUserId);
+    private Param toParam(Long spaceId, UUID currentUserId) {
+        return new Param(spaceId, currentUserId);
     }
 }
