+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/spaceuseraccess/UpdateSpaceLastSeenService.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.users.application.service.space;
+package org.flickit.assessment.users.application.service.spaceuseraccess;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.users.application.port.in.space.UpdateSpaceLastSeenUseCase;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.UpdateSpaceLastSeenUseCase;
 import org.flickit.assessment.users.application.port.out.spaceuseraccess.CheckSpaceAccessPort;
 import org.flickit.assessment.users.application.port.out.spaceuseraccess.UpdateSpaceLastSeenPort;
 import org.springframework.stereotype.Service;
@@ -22,9 +22,9 @@ public class UpdateSpaceLastSeenService implements UpdateSpaceLastSeenUseCase {
 
     @Override
     public void updateLastSeen(Param param) {
-        if (!checkSpaceAccessPort.checkIsMember(param.getId(), param.getCurrentUserId()))
+        if (!checkSpaceAccessPort.checkIsMember(param.getSpaceId(), param.getCurrentUserId()))
             throw new AccessDeniedException(COMMON_CURRENT_USER_NOT_ALLOWED);
 
-        updateSpaceLastSeenPort.updateLastSeen(param.getId(), param.getCurrentUserId(), LocalDateTime.now());
+        updateSpaceLastSeenPort.updateLastSeen(param.getSpaceId(), param.getCurrentUserId(), LocalDateTime.now());
     }
 }
