+++ flickit-assessment-users/src/test/java/org/flickit/assessment/users/application/service/spaceuseraccess/UpdateSpaceLastSeenServiceTest.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.users.application.service.space;
+package org.flickit.assessment.users.application.service.spaceuseraccess;
 
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.users.application.domain.Space;
-import org.flickit.assessment.users.application.port.in.space.UpdateSpaceLastSeenUseCase;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.UpdateSpaceLastSeenUseCase;
 import org.flickit.assessment.users.application.port.out.spaceuseraccess.CheckSpaceAccessPort;
 import org.flickit.assessment.users.application.port.out.spaceuseraccess.UpdateSpaceLastSeenPort;
 import org.flickit.assessment.users.test.fixture.application.SpaceMother;
